refac: organization of files and code of store folder
