+++ store/getters.js
@@ -1,5 +1,7 @@
 export default function getters() {
   return {
-
+    getPokemons(state){
+      return state.pokemons
+    }
   }
 };
